+++ src/Braintree/DisputeGateway.cs
@@ -12,15 +12,15 @@ namespace Braintree
     class DisputeAddEvidenceRequest : Request
     {
         public string Comments { get; set; }
-        public string DocumentUploadId { get; set; }
+        public string DocumentId { get; set; }
 
         public override string ToXml()
         {
             if (Comments != null) {
                 return RequestBuilder.BuildXMLElement("comments", Comments);
             }
-            if (DocumentUploadId != null) {
-                return RequestBuilder.BuildXMLElement("document-upload-id", DocumentUploadId);
+            if (DocumentId != null) {
+                return RequestBuilder.BuildXMLElement("document-upload-id", DocumentId);
             }
             return "";
         }
@@ -33,7 +33,7 @@ namespace Braintree
     {
         private readonly BraintreeService Service;
         private readonly IBraintreeGateway Gateway;
-        private DisputeSearchRequest DisputeSearch { get; set; }        
+        private DisputeSearchRequest DisputeSearch { get; set; }
 
         protected internal DisputeGateway(IBraintreeGateway gateway)
         {
@@ -78,7 +78,7 @@ namespace Braintree
             }
         }
 
-        public virtual Result<DisputeEvidence> AddFileEvidence(string disputeId, string documentUploadId)
+        public virtual Result<DisputeEvidence> AddFileEvidence(string disputeId, FileEvidenceRequest request)
         {
             NotFoundException notFoundException = new NotFoundException(String.Format("dispute with id '{0}' not found", disputeId));
 
@@ -87,13 +87,10 @@ namespace Braintree
                 throw notFoundException;
             }
 
-            if (documentUploadId == null || documentUploadId.Trim().Equals(""))
+            if (request.DocumentId == null || request.DocumentId.Trim().Equals(""))
             {
-                throw new NotFoundException(String.Format("document with id '{0}' not found", documentUploadId));
+                throw new NotFoundException(String.Format("document with id '{0}' not found", request.DocumentId));
             }
-
-            DisputeAddEvidenceRequest request = new DisputeAddEvidenceRequest();
-            request.DocumentUploadId = documentUploadId;
 
             try {
                 XmlNode disputeEvidenceXML = Service.Post(Service.MerchantPath() + "/disputes/" + disputeId + "/evidence", request);
@@ -104,7 +101,12 @@ namespace Braintree
             }
         }
 
-        public virtual async Task<Result<DisputeEvidence>> AddFileEvidenceAsync(string disputeId, string documentUploadId)
+        public virtual Result<DisputeEvidence> AddFileEvidence(string disputeId, string documentUploadId)
+        {
+            return AddFileEvidence(disputeId, new FileEvidenceRequest { DocumentId = documentUploadId });
+        }
+
+        public virtual async Task<Result<DisputeEvidence>> AddFileEvidenceAsync(string disputeId, FileEvidenceRequest request)
         {
             NotFoundException notFoundException = new NotFoundException(String.Format("dispute with id '{0}' not found", disputeId));
 
@@ -113,13 +115,10 @@ namespace Braintree
                 throw notFoundException;
             }
 
-            if (documentUploadId == null || documentUploadId.Trim().Equals(""))
+            if (request.DocumentId == null || request.DocumentId.Trim().Equals(""))
             {
-                throw new NotFoundException(String.Format("document with id '{0}' not found", documentUploadId));
+                throw new NotFoundException(String.Format("document with id '{0}' not found", request.DocumentId));
             }
-
-            DisputeAddEvidenceRequest request = new DisputeAddEvidenceRequest();
-            request.DocumentUploadId = documentUploadId;
 
             try {
                 XmlNode disputeEvidenceXML = await Service.PostAsync(Service.MerchantPath() + "/disputes/" + disputeId + "/evidence", request).ConfigureAwait(false);
@@ -130,26 +129,41 @@ namespace Braintree
             }
         }
 
+        public virtual async Task<Result<DisputeEvidence>> AddFileEvidenceAsync(string disputeId, string documentUploadId)
+        {
+            return await AddFileEvidenceAsync(disputeId, new FileEvidenceRequest { DocumentId = documentUploadId });
+        }
+
         public virtual Result<DisputeEvidence> AddTextEvidence(string disputeId, string content)
         {
-            NotFoundException notFoundException = new NotFoundException(String.Format("dispute with id '{0}' not found", disputeId));
+            TextEvidenceRequest textEvidenceRequest = new TextEvidenceRequest
+            {
+                Content = content
+            };
+            return AddTextEvidence(disputeId, textEvidenceRequest);
+        }
+
+        public virtual Result<DisputeEvidence> AddTextEvidence(string disputeId, TextEvidenceRequest textEvidenceRequest)
+        {
+            NotFoundException notFoundException = new NotFoundException(String.Format("Dispute with ID '{0}' not found", disputeId));
 
             if (disputeId == null || disputeId.Trim().Equals(""))
             {
                 throw notFoundException;
             }
-
-            if (content == null || content.Trim().Equals(""))
+            if (textEvidenceRequest.Content == null || textEvidenceRequest.Content.Trim().Equals(""))
             {
-                throw new ArgumentException("content cannot be empty");
+                throw new ArgumentException("Content cannot be empty");
             }
 
-
-            DisputeAddEvidenceRequest request = new DisputeAddEvidenceRequest();
-            request.Comments = content;
+            int temp;
+            if (textEvidenceRequest.SequenceNumber != null && !int.TryParse(textEvidenceRequest.SequenceNumber, out temp))
+            {
+                throw new ArgumentException("SequenceNumber must be an integer");
+            }
 
             try {
-                XmlNode disputeEvidenceXML = Service.Post(Service.MerchantPath() + "/disputes/" + disputeId + "/evidence", request);
+                XmlNode disputeEvidenceXML = Service.Post(Service.MerchantPath() + "/disputes/" + disputeId + "/evidence", textEvidenceRequest);
 
                 return new ResultImpl<DisputeEvidence>(new NodeWrapper(disputeEvidenceXML), Gateway);
             } catch (NotFoundException) {
@@ -159,23 +173,34 @@ namespace Braintree
 
         public virtual async Task<Result<DisputeEvidence>> AddTextEvidenceAsync(string disputeId, string content)
         {
-            NotFoundException notFoundException = new NotFoundException(String.Format("dispute with id '{0}' not found", disputeId));
+            TextEvidenceRequest textEvidenceRequest = new TextEvidenceRequest
+            {
+                Content = content
+            };
+            return await AddTextEvidenceAsync(disputeId, textEvidenceRequest);
+        }
+
+        public virtual async Task<Result<DisputeEvidence>> AddTextEvidenceAsync(string disputeId, TextEvidenceRequest textEvidenceRequest)
+        {
+            NotFoundException notFoundException = new NotFoundException(String.Format("Dispute with ID '{0}' not found", disputeId));
 
             if (disputeId == null || disputeId.Trim().Equals(""))
             {
                 throw notFoundException;
             }
-
-            if (content == null || content.Trim().Equals(""))
+            if (textEvidenceRequest.Content == null || textEvidenceRequest.Content.Trim().Equals(""))
             {
-                throw new ArgumentException("content cannot be empty");
+                throw new ArgumentException("Content cannot be empty");
             }
 
-            DisputeAddEvidenceRequest request = new DisputeAddEvidenceRequest();
-            request.Comments = content;
+            int temp;
+            if (textEvidenceRequest.SequenceNumber != null && !int.TryParse(textEvidenceRequest.SequenceNumber, out temp))
+            {
+                throw new ArgumentException("SequenceNumber must be an integer");
+            }
 
             try {
-                XmlNode disputeEvidenceXML = await Service.PostAsync(Service.MerchantPath() + "/disputes/" + disputeId + "/evidence", request);
+                XmlNode disputeEvidenceXML = await Service.PostAsync(Service.MerchantPath() + "/disputes/" + disputeId + "/evidence", textEvidenceRequest);
 
                 return new ResultImpl<DisputeEvidence>(new NodeWrapper(disputeEvidenceXML), Gateway);
             } catch (NotFoundException) {