+++ src/Braintree/IOAuthGateway.cs
@@ -4,7 +4,6 @@ namespace Braintree
 {
     public interface IOAuthGateway
     {
-        string ComputeSignature(string message);
         string ConnectUrl(OAuthConnectUrlRequest request);
         ResultImpl<OAuthCredentials> CreateTokenFromCode(OAuthCredentialsRequest request);
         ResultImpl<OAuthCredentials> CreateTokenFromRefreshToken(OAuthCredentialsRequest request);