+++ src/Braintree/TransactionPayPalRequest.cs
@@ -2,6 +2,7 @@ namespace Braintree
 {
     public class TransactionPayPalRequest : Request
     {
+        public string PayeeId { get; set; }
         public string PayeeEmail { get; set; }
 
         public override string ToXml()
@@ -26,7 +27,10 @@ namespace Braintree
 
         protected virtual RequestBuilder BuildRequest(string root)
         {
-            return new RequestBuilder(root).AddElement("payee-email", PayeeEmail);
+            var builder = new RequestBuilder(root);
+            builder.AddElement("payee-id", PayeeId);
+            builder.AddElement("payee-email", PayeeEmail);
+            return builder;
         }
     }
 }