+++ src/Braintree/WebhookTestingGateway.cs
@@ -16,19 +16,23 @@ namespace Braintree
             service = new BraintreeService(gateway.Configuration);
         }
 
-        public virtual Dictionary<string, string> SampleNotification(WebhookKind kind, string id)
+        public virtual Dictionary<string, string> SampleNotification(WebhookKind kind, string id, string sourceMerchantId = null)
         {
             var response = new Dictionary<string, string>();
-            string payload = BuildPayload(kind, id);
+            string payload = BuildPayload(kind, id, sourceMerchantId);
             response["bt_payload"] = payload;
             response["bt_signature"] = BuildSignature(payload);
             return response;
         }
 
-        private string BuildPayload(WebhookKind kind, string id)
+        private string BuildPayload(WebhookKind kind, string id, string sourceMerchantId)
         {
             var currentTime = DateTime.Now.ToUniversalTime().ToString("u");
-            var payload = string.Format("<notification><timestamp type=\"datetime\">{0}</timestamp><kind>{1}</kind><subject>{2}</subject></notification>", currentTime, kind, SubjectSampleXml(kind, id));
+            var sourceMerchantIdXml = "";
+            if (sourceMerchantId != null) {
+                sourceMerchantIdXml = string.Format("<source-merchant-id>{0}</source-merchant-id>", sourceMerchantId);
+            }
+            var payload = string.Format("<notification><timestamp type=\"datetime\">{0}</timestamp><kind>{1}</kind>{2}<subject>{3}</subject></notification>", currentTime, kind, sourceMerchantIdXml, SubjectSampleXml(kind, id));
             return Convert.ToBase64String(Encoding.GetEncoding(0).GetBytes(payload)) + '\n';
         }
 
@@ -63,6 +67,8 @@ namespace Braintree
                 return ConnectedMerchantPayPalStatusChangedSampleXml(id);
             } else if (kind == WebhookKind.PARTNER_MERCHANT_DECLINED) {
                 return PartnerMerchantDeclinedSampleXml(id);
+            } else if (kind == WebhookKind.OAUTH_ACCESS_REVOKED) {
+                return OAuthAccessRevokedSampleXml(id);
             } else if (kind == WebhookKind.DISPUTE_OPENED) {
                 return DisputeOpenedSampleXml(id);
             } else if (kind == WebhookKind.DISPUTE_LOST) {
@@ -235,6 +241,8 @@ namespace Braintree
             return Node("dispute",
                     Node("id", id),
                     Node("amount", "250.00"),
+                    Node("amount-disputed", "250.00"),
+                    Node("amount-won", "245.00"),
                     NodeAttr("received-date", TYPE_DATE, "2014-03-21"),
                     NodeAttr("reply-by-date", TYPE_DATE, "2014-03-21"),
                     Node("currency-iso-code", "USD"),
@@ -253,6 +261,8 @@ namespace Braintree
             return Node("dispute",
                     Node("id", id),
                     Node("amount", "250.00"),
+                    Node("amount-disputed", "250.00"),
+                    Node("amount-won", "245.00"),
                     NodeAttr("received-date", TYPE_DATE, "2014-03-21"),
                     NodeAttr("reply-by-date", TYPE_DATE, "2014-03-21"),
                     Node("currency-iso-code", "USD"),
@@ -271,6 +281,8 @@ namespace Braintree
             return Node("dispute",
                     Node("id", id),
                     Node("amount", "250.00"),
+                    Node("amount-disputed", "250.00"),
+                    Node("amount-won", "245.00"),
                     NodeAttr("received-date", TYPE_DATE, "2014-03-21"),
                     NodeAttr("reply-by-date", TYPE_DATE, "2014-03-21"),
                     Node("currency-iso-code", "USD"),
@@ -373,6 +385,13 @@ namespace Braintree
         private string PartnerMerchantDeclinedSampleXml(string id) {
             return Node("partner-merchant",
                     Node("partner-merchant-id", "abc123")
+            );
+        }
+
+        private string OAuthAccessRevokedSampleXml(string id) {
+            return Node("oauth-application-revocation",
+                    Node("merchant-id", id),
+                    Node("oauth-application-client-id", "oauth_application_client_id")
             );
         }